[Fix #2734] Remove OpenSSL link dependency for osquery core
Move hash table out of utility group, move hash code from core to table
implementation, and replace relevant hash function calls throughout codebase
with appropriate SQL::select...() calls.
