+++ osquery/config/config.cpp
@@ -19,10 +19,10 @@
 #include <osquery/database.h>
 #include <osquery/filesystem.h>
 #include <osquery/flags.h>
-#include <osquery/hash.h>
 #include <osquery/logger.h>
 #include <osquery/packs.h>
 #include <osquery/registry.h>
+#include <osquery/sha1.h>
 #include <osquery/system.h>
 #include <osquery/tables.h>
 
@@ -716,11 +716,10 @@ void Config::getPerformanceStats(
 
 void Config::hashSource(const std::string& source, const std::string& content) {
   WriteLock wlock(config_hash_mutex_);
-  hash_[source] =
-      hashFromBuffer(HASH_TYPE_MD5, &(content.c_str())[0], content.size());
+  hash_[source] = getBufferSHA1(&(content.c_str())[0], content.size());
 }
 
-Status Config::getMD5(std::string& hash) {
+Status Config::getSHA1(std::string& hash) {
   if (!valid_) {
     return Status(1, "Current config is not valid");
   }
@@ -736,8 +735,8 @@ Status Config::getMD5(std::string& hash) {
   for (const auto& it : hash_) {
     add(it.second);
   }
+  hash = getBufferSHA1(&buffer[0], buffer.size());
 
-  hash = hashFromBuffer(HASH_TYPE_MD5, &buffer[0], buffer.size());
   return Status(0, "OK");
 }
