+++ osquery/core/CMakeLists.txt
@@ -15,9 +15,9 @@ ADD_OSQUERY_LIBRARY(TRUE osquery_core
   ${OS_CORE_SOURCE}
   tables.cpp
   flags.cpp
-  hash.cpp
   watcher.cpp
   process_shared.cpp
+  sha1.cpp
 )
 
 file(GLOB OSQUERY_CORE_TESTS "tests/*.cpp")
