+++ osquery/tables/system/darwin/acpi_tables.cpp
@@ -11,7 +11,7 @@
 #include <CoreFoundation/CoreFoundation.h>
 #include <IOKit/IOKitLib.h>
 
-#include <osquery/hash.h>
+#include <osquery/sha1.h>
 #include <osquery/tables.h>
 
 #include "osquery/core/conversions.h"
@@ -22,18 +22,16 @@ namespace tables {
 #define kIOACPIClassName_ "AppleACPIPlatformExpert"
 #define kIOACPIPropertyName_ "ACPI Tables"
 
-void genACPITable(const void *key, const void *value, void *results) {
+void genACPITable(const void* key, const void* value, void* results) {
   Row r;
-
-  r["name"] = stringFromCFString((CFStringRef)key);
-
   auto data = (CFDataRef)value;
   auto length = CFDataGetLength(data);
-  r["size"] = INTEGER(length);
-  r["md5"] =
-      osquery::hashFromBuffer(HASH_TYPE_MD5, CFDataGetBytePtr(data), length);
 
-  ((QueryData *)results)->push_back(r);
+  r["name"] = stringFromCFString((CFStringRef)key);
+  r["size"] = INTEGER(length);
+  r["sha1"] = getBufferSHA1(CFDataGetBytePtr(data), length);
+
+  ((QueryData*)results)->push_back(r);
 }
 
 QueryData genACPITables(QueryContext& context) {
