+++ osquery/tables/system/hash.cpp
@@ -12,9 +12,12 @@
 #include <sstream>
 #include <vector>
 
+#include <boost/filesystem.hpp>
+
 #include <osquery/filesystem.h>
 #include <osquery/hash.h>
 #include <osquery/logger.h>
+#include <osquery/tables.h>
 
 namespace osquery {
 
@@ -22,8 +25,8 @@ namespace osquery {
 #import <CommonCrypto/CommonDigest.h>
 #define __HASH_API(name) CC_##name
 #else
-#include <openssl/sha.h>
 #include <openssl/md5.h>
+#include <openssl/sha.h>
 #define __HASH_API(name) name
 
 #define SHA1_DIGEST_LENGTH SHA_DIGEST_LENGTH
@@ -138,5 +141,104 @@ std::string hashFromFile(HashType hash_type, const std::string& path) {
   } else {
     return hashes.sha256;
   }
+}
+
+namespace tables {
+
+void genHashForFile(const std::string& path,
+                    const std::string& dir,
+                    QueryContext& context,
+                    QueryData& results) {
+  // Must provide the path, filename, directory separate from boost path->string
+  // helpers to match any explicit (query-parsed) predicate constraints.
+  Row r;
+  if (context.isCached(path)) {
+    r = context.getCache(path);
+  } else {
+    auto hashes = hashMultiFromFile(
+        HASH_TYPE_MD5 | HASH_TYPE_SHA1 | HASH_TYPE_SHA256, path);
+
+    r["path"] = path;
+    r["directory"] = dir;
+    r["md5"] = std::move(hashes.md5);
+    r["sha1"] = std::move(hashes.sha1);
+    r["sha256"] = std::move(hashes.sha256);
+    context.setCache(path, r);
+  }
+  results.push_back(r);
+}
+
+QueryData genHash(QueryContext& context) {
+  QueryData results;
+  boost::system::error_code ec;
+
+  // The query must provide a predicate with constraints including path or
+  // directory. We search for the parsed predicate constraints with the equals
+  // operator.
+  auto paths = context.constraints["path"].getAll(EQUALS);
+  context.expandConstraints(
+      "path",
+      LIKE,
+      paths,
+      ([&](const std::string& pattern, std::set<std::string>& out) {
+        std::vector<std::string> patterns;
+        auto status =
+            resolveFilePattern(pattern, patterns, GLOB_ALL | GLOB_NO_CANON);
+        if (status.ok()) {
+          for (const auto& resolved : patterns) {
+            out.insert(resolved);
+          }
+        }
+        return status;
+      }));
+
+  // Iterate through the file paths, adding the hash results
+  for (const auto& path_string : paths) {
+    boost::filesystem::path path = path_string;
+    if (!boost::filesystem::is_regular_file(path, ec)) {
+      continue;
+    }
+
+    genHashForFile(path_string, path.parent_path().string(), context, results);
+  }
+
+  // Now loop through constraints using the directory column constraint.
+  auto directories = context.constraints["directory"].getAll(EQUALS);
+  context.expandConstraints(
+      "directory",
+      LIKE,
+      directories,
+      ([&](const std::string& pattern, std::set<std::string>& out) {
+        std::vector<std::string> patterns;
+        auto status =
+            resolveFilePattern(pattern, patterns, GLOB_FOLDERS | GLOB_NO_CANON);
+        if (status.ok()) {
+          for (const auto& resolved : patterns) {
+            out.insert(resolved);
+          }
+        }
+        return status;
+      }));
+
+  // Iterate over the directory paths
+  for (const auto& directory_string : directories) {
+    boost::filesystem::path directory = directory_string;
+    if (!boost::filesystem::is_directory(directory, ec)) {
+      continue;
+    }
+
+    // Iterate over the directory files and generate a hash for each regular
+    // file.
+    boost::filesystem::directory_iterator begin(directory), end;
+    for (; begin != end; ++begin) {
+      if (boost::filesystem::is_regular_file(begin->path(), ec)) {
+        genHashForFile(
+            begin->path().string(), directory_string, context, results);
+      }
+    }
+  }
+
+  return results;
+}
 }
 }
